+++ database/sqlite_test.go
@@ -33,27 +33,3 @@ func TestSqlite(t *testing.T) {
 
 	assert.Equal(t, meaning, "Ivory black; animal charcoal.")
 }
-
-func TestBadger(t *testing.T) {
-	t.Parallel()
-
-	tmpFile := testhelper.GetTempFileLoc()
-
-	h, err := database.NewBadger(tmpFile + "/")
-	require.NoError(t, err, "creating handler")
-
-	t.Cleanup(func() {
-		require.NoError(t, h.Close(), "closing db")
-	})
-
-	err = h.Insert("abiser", "Ivory black; animal charcoal.")
-	require.NoError(t, err, "inserting value")
-
-	err = h.Insert("abiser", "Ivory black; animal charcoal.")
-	require.NoError(t, err, "inserting value second time")
-
-	meaning, err := h.Get("abiser")
-	require.NoError(t, err, "getting value")
-
-	assert.Equal(t, meaning, "Ivory black; animal charcoal.")
-}
